Added project to Github repository
I did not store my progress for this project in a Github repository and instead stored in locally within my computer. It was not until now that I actually stored my project in a repository. If there are any issues with the project not working due to missing or corrupt files, I apologize for the inconvenience.